check overflow of multiplication

+++ include/util/random.h
@@ -7,7 +7,7 @@
 // perm[0...k-1] is the results
 // but assume `perm` has allocated n spaces
 
-void rand_perm (uint64_t *perm, uint64_t n, uint64_t k, int32_t seed);
+void rand_perm (int64_t *perm, int64_t n, int64_t k, int64_t seed);
 float rand_float();
 int rand_int();
